rename qiskit utils
